EE-9537 Move special character name processing into dedicated component
