+++ src/main/java/uk/gov/digital/ho/pttg/application/namematching/NameMatchingCandidatesService.java
@@ -1,26 +1,27 @@
 package uk.gov.digital.ho.pttg.application.namematching;
 
-import com.google.common.collect.Lists;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import uk.gov.digital.ho.pttg.application.namematching.candidates.MultipleLastNames;
 import uk.gov.digital.ho.pttg.application.namematching.candidates.NameCombinations;
+import uk.gov.digital.ho.pttg.application.namematching.candidates.SpecialCharacters;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
 
 import static uk.gov.digital.ho.pttg.application.namematching.NameMatchingFunctions.deduplicate;
 
 @Service
 public class NameMatchingCandidatesService {
-    private static final String NAME_SPLITTERS = "-'";
-    private static final String NAME_SPLITTER_REGEX = "[" + NAME_SPLITTERS + "]";
 
     private NameCombinations nameCombinations;
     private MultipleLastNames multipleLastNames;
+    private SpecialCharacters specialCharacters;
 
-    public NameMatchingCandidatesService(NameCombinations nameCombinations, MultipleLastNames multipleLastNames) {
+    public NameMatchingCandidatesService(NameCombinations nameCombinations, MultipleLastNames multipleLastNames, SpecialCharacters specialCharacters) {
         this.nameCombinations = nameCombinations;
         this.multipleLastNames = multipleLastNames;
+        this.specialCharacters = specialCharacters;
     }
 
     public List<PersonName> generateCandidateNames(String firstName, String lastName) {
@@ -29,36 +30,9 @@ public class NameMatchingCandidatesService {
 
         candidates.addAll(multipleLastNames.generateCandidates(firstName, lastName));
         candidates.addAll(nameCombinations.generateCandidates(firstName, lastName));
-
-        if (namesContainSplitters(firstName, lastName)) {
-            candidates.addAll(generateCandidatesWithSplitters(firstName, lastName));
-        }
+        candidates.addAll(specialCharacters.generateCandidates(firstName, lastName));
 
         return Collections.unmodifiableList(deduplicate(candidates));
     }
-
-    private static boolean namesContainSplitters(String firstName, String lastName) {
-        return StringUtils.containsAny(firstName, NAME_SPLITTERS) || StringUtils.containsAny(lastName, NAME_SPLITTERS);
-    }
-
-    private List<PersonName> generateCandidatesWithSplitters(String firstName, String lastName) {
-        Set<PersonName> candidateNames = new LinkedHashSet<>();
-
-        candidateNames.addAll(multipleLastNames.generateCandidates(nameWithSplittersRemoved(firstName), nameWithSplittersRemoved(lastName)));
-        candidateNames.addAll(multipleLastNames.generateCandidates(nameWithSplittersReplacedBySpaces(firstName), nameWithSplittersReplacedBySpaces(lastName)));
-        candidateNames.addAll(nameCombinations.generateCandidates(nameWithSplittersRemoved(firstName), nameWithSplittersRemoved(lastName)));
-        candidateNames.addAll(nameCombinations.generateCandidates(nameWithSplittersReplacedBySpaces(firstName), nameWithSplittersReplacedBySpaces(lastName)));
-
-        return Lists.newArrayList(candidateNames);
-    }
-
-    private static String nameWithSplittersRemoved(String name) {
-        return name.replaceAll(NAME_SPLITTER_REGEX, "");
-    }
-
-    private static String nameWithSplittersReplacedBySpaces(String name) {
-        return name.replaceAll(NAME_SPLITTER_REGEX, " ");
-    }
-
 
 }
